refactor(codehealth): tools package to green
